+++ CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/translators/ScoreToTableTranslator.java
@@ -48,15 +48,11 @@ public class ScoreToTableTranslator extends AbstractTableTranslator<ScoreResult,
         final TableVisualisation<ElementScore> tableVisualisation = new TableVisualisation<>(this);
         final Set<String> scoreNames = result.getUniqueScoreNames();
         tableVisualisation.addColumn(IDENTIFIER_COLUMN_NAME, (100 / (scoreNames.size() + 2)) * 2);
-        scoreNames.forEach(scoreName -> {
-            tableVisualisation.addColumn(scoreName, (100 / (scoreNames.size() + 2)));
-        });
+        scoreNames.forEach(scoreName -> tableVisualisation.addColumn(scoreName, (100 / (scoreNames.size() + 2))));
         tableVisualisation.populateTable(result.isIgnoreNullResults()
                 ? result.get().stream().filter(elementMultiScore -> !elementMultiScore.isNull()).collect(Collectors.toList()) : result.get());
         result.addResultListener(tableVisualisation);
-        tableVisualisation.setSelectionModelListener(change -> {
-            result.setSelectionOnGraph(tableVisualisation.getSelectedItems());
-        });
+        tableVisualisation.setSelectionModelListener(change -> result.setSelectionOnGraph(tableVisualisation.getSelectedItems()));
         return tableVisualisation;
     }